+++ Modules/Sources/PocketCastsServer/Public/Search/PodcastSearchTask.swift
@@ -24,6 +24,11 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
     public let author: String?
     public let kind: Kind
     public var isLocal: Bool?
+    public var explicit: Bool?
+
+    enum CodingKeys: String, CodingKey {
+        case uuid, title, author, kind, isLocal, explicit
+    }
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
@@ -32,6 +37,7 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
         self.author = try? container.decode(String.self, forKey: .author)
         self.kind = (try? container.decodeIfPresent(Kind.self, forKey: .kind)) ?? .podcast
         self.isLocal = (try? container.decode(Bool.self, forKey: .isLocal)) ?? false
+        self.explicit = try? container.decodeIfPresent(Bool.self, forKey: .explicit)
     }
 
     public init?(from podcast: Podcast) {
@@ -40,6 +46,7 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
         self.author = podcast.author
         self.isLocal = true
         self.kind = .podcast
+        self.explicit = podcast.isExplicit
     }
 
     public init?(from folder: Folder) {
@@ -48,6 +55,7 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
         self.author = ""
         self.isLocal = true
         self.kind = .folder
+        self.explicit = false
     }
 
     public init?(from predictiveResult: PredictiveSearchResult) {
@@ -58,6 +66,7 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
                 self.title = podcast.title
                 self.kind = .podcast
                 self.isLocal = false
+                self.explicit = podcast.isExplicit
             default:
                 return nil
         }
@@ -72,6 +81,7 @@ public struct PodcastFolderSearchResult: Codable, Hashable {
         self.title = combinedResult.title
         self.kind = .podcast
         self.isLocal = false
+        self.explicit = combinedResult.explicit
     }
 
     public enum Kind: Codable {