+++ Pocket Casts TV App/Data/TVDataManager.swift
@@ -62,4 +62,34 @@ class TVDataManager {
         PlaybackActionHelper.play(episode: episode, podcastUuid: podcastUuid)
         return true
     }
+
+    func playEpisode(_ episode: DiscoverEpisode) async -> Bool {
+        guard let podcastUuid = episode.podcastUuid, let episodeUuid = episode.uuid else {
+            return false
+        }
+        return await playEpisode(podcastUuid: podcastUuid, episodeUuid: episodeUuid)
+    }
+
+    func playEpisode(podcastUuid: String, episodeUuid: String) async -> Bool {
+        _ = await loadPodcast(podcastUuid: podcastUuid)
+
+        guard let episode = dataManager.findEpisode(uuid: episodeUuid) else {
+            return false
+        }
+
+        //if we are already playing the episode let's return true
+        guard !playbackManager.isActivelyPlaying(episodeUuid: episode.uuid) else {
+            return true
+        }
+
+        PlaybackActionHelper.play(episode: episode, podcastUuid: podcastUuid)
+        return true
+    }
+
+    func playEpisode(_ episode: EpisodeSearchResult) async -> Bool {
+        guard !episode.podcastUuid.isEmpty, !episode.uuid.isEmpty else {
+            return false
+        }
+        return await playEpisode(podcastUuid: episode.podcastUuid, episodeUuid: episode.uuid)
+    }
 }