+++ podcasts/PaidFeature.swift
@@ -9,6 +9,7 @@ extension PaidFeature {
     static var bookmarks: PaidFeature = .plusFeature
     static var deselectChapters: PaidFeature = .plusFeature
     static var slumber: PaidFeature = .plusFeature
+    static var syncedTranscripts: PaidFeature = .plusFeature
 }
 
 /// A `PaidFeature` represents a feature that is unlocked with a subscription tier, and is considered to be unlocked if the tier